wip #616 open stats
